# project02/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is an api server for a recipe keeper website.\n",
    "version": "1.0.0",
    "title": "CSE 341 Project 02 - Recipe Keeper",
    "contact": {
      "email": "rosnikitin90@gmail.com"
    }
  },
  "host": "cse-341-w06-nikitin-r.onrender.com",
  "basePath": "/",
  "tags": [
    {
      "name": "recipe",
      "description": "Operations about recipes"
    },
    {
      "name": "user",
      "description": "Operations about user"
    }
  ],
  "schemes": ["https"],
  "paths": {
    "/recipe/{recipeName}": {
      "get": {
        "tags": ["recipe"],
        "summary": "Finds Recipe by name",
        "description": "Each recipe has a unique name, which serves as a unique id, making it possible to find a single recipe by the exact name.",
        "operationId": "findRecipeByName",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "recipeName",
            "in": "path",
            "description": "Recipe name string is required and functions as id",
            "required": true,
            "type": "string",
            "items": {
              "type": "string",
              "default": "available"
            },
            "collectionFormat": "csv"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object"
            }
          },
          "400": {
            "description": "Invalid recipe name value"
          }
        }
      }
    },
    "/user/{username}": {
      "get": {
        "tags": ["user"],
        "summary": "Get user by user name",
        "operationId": "getUserByName",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "The name that needs to be fetched.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Invalid username supplied"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/user": {
      "get": {
        "tags": ["user"],
        "summary": "Get all users",
        "operationId": "getAllUsers",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "Array",
              "$ref": "#/definitions/User"
            }
          },
          "500": {
            "description": "Error occurred"
          }
        }
      },
      "post": {
        "tags": ["user"],
        "summary": "Create user",
        "description": "This is for when a user creates an account.",
        "operationId": "createUser",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Created user object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Credentials"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "successfully created user",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer",
                  "format": "int64"
                },
                "name": {
                  "type": "string",
                  "example": "Recipe Name"
                }
              }
            }
          },
          "400": {
            "description": "Invalid theme name value"
          }
        }
      },
      "put": {
        "tags": ["user"],
        "summary": "Update user",
        "description": "",
        "operationId": "updateUser",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "The name to be updated.",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Updated user object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successfully modified user"
          },
          "400": {
            "description": "Invalid user supplied"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "delete": {
        "tags": ["user"],
        "summary": "Delete user",
        "description": "",
        "operationId": "deleteUser",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "The name to be deleted",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "Successfully deleted user"
          },
          "400": {
            "description": "Invalid username supplied"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "required": ["password", "username"],
      "properties": {
        "_id": {
          "type": "integer",
          "format": "string"
        },
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "displayName": {
          "type": "string"
        },
        "info": {
          "$ref": "#/definitions/User_info"
        },
        "profile": {
          "$ref": "#/definitions/User_profile"
        }
      }
    },
    "Recipe": {
      "type": "object",
      "required": ["ingredients", "name", "preparationTime", "servings", "steps"],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string",
          "example": "concise legible recipe name"
        },
        "servings": {
          "type": "integer",
          "example": 18
        },
        "preparationTime": {
          "type": "string",
          "example": "10 min"
        },
        "ingredients": {
          "type": "array",
          "example": ["", "Ingredient Number One...", "Ingredient Number Two...", "Ingredient Number Three...", "Ingredient Number Four...", "Ingredient Number Five..."],
          "items": {
            "type": "string"
          }
        },
        "steps": {
          "type": "array",
          "example": ["", "Description of Step One...", "Description of Step Two...", "Description of Step Three...", "Description of Step Four...", "Description of Step Five..."],
          "items": {
            "type": "string"
          }
        }
      }
    },
    "Credentials": {
      "type": "object",
      "required": ["password", "username"],
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "User_info": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "currentLocation": {
          "type": "string"
        },
        "openToNewOpportunities": {
          "type": "boolean"
        },
        "profileIsPublic": {
          "type": "boolean"
        },
        "recipe_name": {
          "type": "string"
        }
      }
    },
    "User_profile_experience": {
      "type": "object",
      "properties": {
        "jobTitle": {
          "type": "string"
        },
        "jobDescription": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "dateStart": {
          "type": "string"
        },
        "dateEnd": {
          "type": "string"
        },
        "company": {
          "type": "string"
        },
        "location": {
          "type": "string"
        }
      }
    },
    "User_profile_education": {
      "type": "object",
      "properties": {
        "institution": {
          "type": "string"
        },
        "programTitle": {
          "type": "string"
        },
        "programDescription": {
          "type": "string"
        },
        "dateStart": {
          "type": "string"
        },
        "dateEnd": {
          "type": "string"
        },
        "completed": {
          "type": "boolean"
        },
        "location": {
          "type": "string"
        },
        "topics": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "User_profile_projects": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "dateStart": {
          "type": "string"
        },
        "dateEnd": {
          "type": "string"
        },
        "githubUrl": {
          "type": "string"
        },
        "hostedUrl": {
          "type": "string"
        },
        "images": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "skills": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "User_profile": {
      "type": "object",
      "properties": {
        "experience": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/User_profile_experience"
          }
        },
        "education": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/User_profile_education"
          }
        },
        "projects": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/User_profile_projects"
          }
        },
        "skills": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    }
  }
}